requirements.txt
streamlit
google-genai
requests
pandas
python-dateutil
tenacity